feat: model add for tables and files

+++ app/backend/traits/Service.php
@@ -186,6 +186,7 @@ trait Service
             // handle descendant
             $tree = $this->treeDataAPI(['trash' => 'withTrashed']);
             $allIds = [];
+            $body = [];
             foreach ($ids as $id) {
                 $allIds[] = (int)$id;
                 $allIds = array_merge($allIds, getDescendantSet('id', 'id', $id, $tree));
@@ -193,16 +194,21 @@ trait Service
 
             if ($type === 'deletePermanently') {
                 $dataSet = $this->withTrashed()->whereIn('id', array_unique($allIds))->select();
-                foreach ($dataSet as $item) {
-                    $item->force()->delete();
+                if (!$dataSet->isEmpty()) {
+                    $body = $dataSet->toArray();
+                    foreach ($dataSet as $item) {
+                        $item->force()->delete();
+                    }
+                    $result = true;
+                } else {
+                    return $this->error('Nothing to do.');
                 }
-                $result = true;
             } else {
                 $result = $this->withTrashed()->whereIn('id', array_unique($allIds))->select()->delete();
             }
-
+            
             if ($result) {
-                return $this->success('Delete successfully.');
+                return $this->success('Delete successfully.', $body);
             } else {
                 return $this->error('Delete failed.');
             }
